refactor(core): modularize scene handling and cleanup logic
Moved scene initialization and cleanup logic into dedicated functions (`run_scene` and `clean_scene`) to improve code modularity and readability. Replaced `ssize_t` with `int` for file descriptors for better type consistency. Added a new error message for MLX initialization failure.

+++ sources/check_file.c
@@ -14,7 +14,7 @@
 
 void	check_file(char	*file)
 {
-	ssize_t	fd;
+	int	fd;
 
     fd = open(file, O_DIRECTORY);
     if (fd >= 0)
@@ -33,7 +33,7 @@ void	check_file(char	*file)
 
 bool	is_valid_png(char *file)
 {
-    ssize_t fd;
+    int	fd;
 
     fd = open(file, O_DIRECTORY);
     if (fd >= 0)
